Refinements

+++ app/src/main/java/com/autocrop/activities/cropping/fragments/cropping/CropFragment.kt
@@ -15,6 +15,7 @@ import com.autocrop.activities.cropping.fragments.croppingfailed.CroppingFailedF
 import com.autocrop.activities.iodetermination.IODeterminationActivity
 import com.autocrop.activities.iodetermination.IODeterminationActivityViewModel
 import com.autocrop.activities.main.MainActivity
+import com.autocrop.utils.android.extensions.getLong
 import com.autocrop.utils.android.extensions.loadBitmap
 import com.autocrop.utils.kotlin.extensions.executeAsyncTaskWithProgressUpdateReceiver
 import com.blogspot.atifsoftwares.animatoolib.Animatoo
@@ -91,7 +92,7 @@ class CropFragment
                         .fragmentReplacementTransaction(CroppingFailedFragment())
                         .commit()
                 },
-                resources.getInteger(R.integer.delay_minimal).toLong()
+                resources.getLong(R.integer.delay_minimal)
             )
     }
