+++ app/src/main/java/com/autocrop/activities/main/fragments/flowfield/views/ButtonsLayout.kt
@@ -6,6 +6,7 @@ import android.widget.RelativeLayout
 import com.autocrop.activities.main.MainActivityViewModel
 import com.autocrop.utils.android.extensions.activityViewModel
 import com.autocrop.utils.android.extensions.fadeIn
+import com.autocrop.utils.android.extensions.getLong
 import com.autocrop.utils.android.extensions.ifNotInEditMode
 import com.autocrop.utils.android.extensions.show
 import com.w2sv.autocrop.R
@@ -19,7 +20,7 @@ class ButtonsLayout(context: Context, attributeSet: AttributeSet):
         ifNotInEditMode {
             with(activityViewModel<MainActivityViewModel>()){
                 if (fadeInFlowFieldButtons){
-                    fadeIn(resources.getInteger(R.integer.duration_flowfield_buttons_fade_in).toLong())
+                    fadeIn(resources.getLong(R.integer.duration_flowfield_buttons_fade_in))
                     fadeInFlowFieldButtons = false
                 }
                 else
